+++ docs/api_domain.md
@@ -3,11 +3,11 @@ hide:
   - navigation
 ---
 
-# Schema References for `mostlyai.client.domain`
+# Schema References for `mostlyai.domain`
 
 This module is auto-generated to represent `pydantic`-based classes of the defined schema in the [Public API](https://github.com/mostly-ai/mostly-openapi/blob/main/public-api.yaml).
 
-::: mostlyai.client.domain
+::: mostlyai.domain
     options:
       show_root_heading: true
       show_root_full_path: true
