+++ mostlyai/client/_mostly_utils.py
@@ -28,7 +28,7 @@ from rich.progress import (
 from rich.style import Style
 
 from mostlyai.client._base_utils import convert_to_base64, read_table_from_path
-from mostlyai.client.domain import (
+from mostlyai.domain import (
     StepCode,
     ProgressStatus,
     Generator,
