+++ mostlyai/client/synthetic_datasets.py
@@ -20,7 +20,7 @@ from typing import Any, Iterator, Optional, Union
 import pandas as pd
 
 from mostlyai.client.base import DELETE, GET, PATCH, POST, Paginator, _MostlyBaseClient
-from mostlyai.client.domain import (
+from mostlyai.domain import (
     JobProgress,
     SyntheticDataset,
     SyntheticDatasetFormat,
